+++ src/main/java/com/codeit/sb02mplteam2/domain/playlist/dto/request/PlaylistUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.codeit.sb02mplteam2.domain.playlist.dto;
+package com.codeit.sb02mplteam2.domain.playlist.dto.request;
 
 public record PlaylistUpdateRequest(
     String newTitle,
