+++ src/main/java/com/codeit/sb02mplteam2/domain/playlist/dto/request/SubscribeRequest.java
@@ -1,4 +1,4 @@
-package com.codeit.sb02mplteam2.domain.playlist.service;
+package com.codeit.sb02mplteam2.domain.playlist.dto.request;
 
 public record SubscribeRequest(
     Long userId,
